Use ToolUtilities
