+++ .changeset/five-icons-sneeze.md
@@ -1,0 +1,5 @@
+---
+"@itwin/appui-react": minor
+---
+
+Use `ToolUtilities.defineIcon` to define icons for `RotateViewTool` and `ViewToggleCameraTool` tools.
